+++ CHANGES.md
@@ -4,6 +4,7 @@ CHANGES
 ## Development
 
 * Port to Python 3, use Unicode in Python 2
+* Modularize Blogger API use, new services for adding new services
 
 ## Version 0.5.2 (2013-07-29T03:37:44Z)